requirements.txt
asgiref==3.3.1
black==21.6b0
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
cryptography
defusedxml==0.6.0
Django==3.2
django-allauth
django-solo==1.2.0
flake8===3.9.2
idna==2.10
isort==5.9.1
oauthlib==3.1.0
pycparser==2.20
pydantic==1.8.2
pymemcache==3.5.0
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2020.4
pytest==6.2.4
pytest-django==4.4.0
requests==2.25.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
urllib3==1.26.2
xmltodict==0.12.0
boto3==1.18.22
redis==3.5.3
celery==4.4.7
flower==0.9.7
